fix : 최종 수익 실수형으로 출력되는 버그 수정
Dealer 클래스와 Player 클래스에 있는 getEarnMoney() 함수의 리턴형을 double 형에서 int 형으로 수정.

+++ src/main/java/domain/user/Dealer.java
@@ -86,12 +86,12 @@ public class Dealer {
         return smallestCard;
     }
 
-    public double getEarnMoney(ArrayList<Player> players, int maxValue) {
+    public int getEarnMoney(ArrayList<Player> players, int maxValue) {
         double dealerMoney = 0;
 
         for (Player player : players) {
             dealerMoney -= player.getEarnMoney(maxValue);
         }
-        return dealerMoney;
+        return (int) dealerMoney;
     }
 }
